Java files: Added ChatMsg.java,ChatMsgAdapter.java
ChatMsg.java will be handling each individual message/image for the
app. ChatMsgAdapter.java is meant to handle the ListView, where the
chats shall be shown.
In ChatMsg.java, the functions are meant to get and set certain
parameters, like content, and whether a message is image or text. In
ChatMsgAdapter.java, these values are fetched or set accordingly.
Constants are kept to maintain status of the message received or sent.

Signed-off-by: Rohit Das <rohit.das950@gmail.com>